Update README.md
Themes and Ideas in Lucifer are talked about a lot in the show. One of the themes that are closely related to the show is Good vs Evil since it often seeks the answer to such questions as redemption, forgiveness, and the nature of good and evil. Lucifer, the Devil himself, ponders on his morality and sometimes even doubts his position in the universe. The second most important theme is Free Will which is the aspect of the inner struggle that Lucifer is going through concerning the idea of destiny and his own right to choose. This is the consequence of the theme of Identity, as he finds it difficult to know who he is, both as an angel and as the Devil, which eventually leads him to look for his real self. Apart from that, the show also looks at Love & Relationships, especially the distinguishing, but also often the complex, connection between Lucifer and Chloe Decker. They are not just coworkers, their personal lives are enmeshed in a way that tests Lucifer’s views on love, trust, and vulnerability. The series is also dealing with the complicated relationships of Lucifer and his celestial family which are the factors that give a new dimension to the story.
Apart from that, some other Impressive Scenes are undoubtedly the series' main strengths. One of the most heavy-hearted moments is when Lucifer makes his First Appear which is the time that he expresses his devilish true form for the first time. On the one hand, its sheer horror aspect is the reason for its awe but on the other, it is the scene that marks a major turning point in the show. Another notable scene is when Lucifer realizes that he possesses a special power that enables him to make people confess their deepest desires. This gift or "the persuasive power" soon becomes a primary tool of his alliance with Chloe as they both use it to bring justice to those guilty of crimes. Among others, these moments are the ones that built the road for the intriguing conversation that the show had with the complex questions of morality, identity, and the human (and celestial) condition.